allow attach prop on Object3D (#1372)
There are cases where you might want to attach an Object3D child to a property of a parent, just like you can with non-Object3D instances today. For example, you might have a render pass (managed in EffectComposer) with a `scene` member, and you want to have its scene in the R3F tree.

This change flips the order that the isObject3D check and the child.attach* checks run in appendChild and insertBefore, so that an Object3D child that has one of the attach props will get attached instead of being a scene-graph child, regardless of whether the parent is an Object3D. It also updates removeChild to match.

The test suite gets a new test for this behaviour.